Issue #53: Pad incomplete voices in MeasureBuilder when building
If there is a voice in a MeasureBuilder that does not contain
enough durational notation elements to fill a measure with the
set time signature, then the voice is padded with rests at
the end.

Add test for checking that voices are padded.

Update other tests to expect the new functionality.