Issue #6: Implement equalsTranpositionally
Refactor other methods to allow checking the
equality of two voices with different transformation
applied to them.

This implementation is quite brute-force, but
should work ok with small patterns. If this
implementation proves too slow with large patterns,
it can be replaced with point-set comparisons.
Another option is to keep the voices sorted lexicographically
by pitch height to avoid having to check more than one transposition.