Remove additional file exists check
The file exists check causes IOException to be thrown
only if file doesn't exist when the exception should
also be thrown if opening file fails due to lacking
permissions.

Also checking first for file properties and then
trying to open it is an antipattern because the
file may have changed between the check and
opening.